Ok, ok, I think I got it now!

+++ src/main/java/JTiled/Tile.java
@@ -23,7 +23,7 @@ public class Tile {
     @XStreamAsAttribute
     int terrian;
     @XStreamAsAttribute
-    int wallFlags;
+    long wallFlags;
 
     Tile(Vector2i pos, int tilesetId) {
         this.pos = pos;
